PIMCORE-1340 - Possibility to define redirects directly in a documents (aliases) =>check if parameter settings is defined, because the data isn't sent if the tab wasn't activated
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@1123 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/modules/admin/controllers/PageController.php
@@ -73,40 +73,42 @@ class Admin_PageController extends Pimcore_Controller_Action_Admin_Document {
             }
 
             // check for redirects
-            if($this->getUser()->isAllowed("redirects")) {
+            if($this->getUser()->isAllowed("redirects") && $this->_getParam("settings")) {
                 $settings = Zend_Json::decode($this->_getParam("settings"));
 
-                $redirectList = new Redirect_List();
-                $redirectList->setCondition("target = ?", $page->getId());
-                $existingRedirects = $redirectList->load();
-                $existingRedirectIds = array();
-                foreach ($existingRedirects as $existingRedirect) {
-                    $existingRedirectIds[$existingRedirect->getId()] = $existingRedirect->getId();
-                }
-
-                for($i=1;$i<100;$i++) {
-                    if(array_key_exists("redirect_url_".$i, $settings)) {
-
-                        // check for existing
-                        if($settings["redirect_id_".$i]) {
-                            $redirect = Redirect::getById($settings["redirect_id_".$i]);
-                            unset($existingRedirectIds[$redirect->getId()]);
-                        } else {
-                            // create new one
-                            $redirect = new Redirect();
-                        }
-
-                        $redirect->setSource($settings["redirect_url_".$i]);
-                        $redirect->setTarget($page->getId());
-                        $redirect->setStatusCode(301);
-                        $redirect->save();
+                if(is_array($settings)) {
+                    $redirectList = new Redirect_List();
+                    $redirectList->setCondition("target = ?", $page->getId());
+                    $existingRedirects = $redirectList->load();
+                    $existingRedirectIds = array();
+                    foreach ($existingRedirects as $existingRedirect) {
+                        $existingRedirectIds[$existingRedirect->getId()] = $existingRedirect->getId();
                     }
-                }
 
-                // remove existing redirects which were delete
-                foreach ($existingRedirectIds as $existingRedirectId) {
-                    $redirect = Redirect::getById($existingRedirectId);
-                    $redirect->delete();
+                    for($i=1;$i<100;$i++) {
+                        if(array_key_exists("redirect_url_".$i, $settings)) {
+
+                            // check for existing
+                            if($settings["redirect_id_".$i]) {
+                                $redirect = Redirect::getById($settings["redirect_id_".$i]);
+                                unset($existingRedirectIds[$redirect->getId()]);
+                            } else {
+                                // create new one
+                                $redirect = new Redirect();
+                            }
+
+                            $redirect->setSource($settings["redirect_url_".$i]);
+                            $redirect->setTarget($page->getId());
+                            $redirect->setStatusCode(301);
+                            $redirect->save();
+                        }
+                    }
+
+                    // remove existing redirects which were delete
+                    foreach ($existingRedirectIds as $existingRedirectId) {
+                        $redirect = Redirect::getById($existingRedirectId);
+                        $redirect->delete();
+                    }
                 }
             }
